리스트 삭제 애니메이션 구현

+++ notice-board/pages/api/deletePost.js
@@ -5,10 +5,13 @@ export default async function hander(req, res) {
   try {
     const client = await connectDB;
     const db = client.db("NextForum");
+    let result = null;
 
-    const result = await db.collection('posts').deleteOne({
-      _id: new ObjectId(req.query.id),
-    });
+    if (req.method == 'DELETE') {
+      result = await db.collection('posts').deleteOne({
+        _id: new ObjectId(req.query.id),
+      });
+    }
 
     res.status(200).json(result);
   } catch (e) {
